requirements.txt
asgiref==3.5.0
boto3==1.20.46
botocore==1.23.46
dj-database-url==0.5.0
Django==4.0.2
django-appconf==1.0.5
django-extensions==3.1.5
django-on-heroku==1.1.2
django-slugger==1.1.2
django-storages==1.12.3
gunicorn==20.1.0
jmespath==0.10.0
pilkit==2.0
Pillow==9.0.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.19.2
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
Unidecode==1.3.2
urllib3==1.26.8
whitenoise==5.3.0
